requirements.txt
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate 
Flask-JWT==0.3.2
PyJWT
psycopg2